CardMatchingGame: Use Deck instead of Pointer
Moved away from having all cards, both in play cards and to be draw cards, in the same cards array. Previously a pointer kept track of which was the next card to be drawn. Now we have a deck for the cards to be drawn and an array of cards with the cards in play.